+++ backend-nestjs/src/products/products.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { ProductsController } from './products.controller';
 import { CategoriesModule } from '../categories/categories.module';
+import { UploadModule } from '../upload/upload.module';
 
 @Module({
-  imports: [CategoriesModule],
+  imports: [CategoriesModule, UploadModule],
   controllers: [ProductsController],
   providers: [ProductsService],
   exports: [ProductsService],